refactor: rename swapUtils.ts to tokenUtils.ts
The file is used across borrowing, lending, pools, orchestrator and swap.
The name swapUtils was misleading — tokenUtils better reflects its scope
as a shared utility module for token/amount operations.
