+++ apps/web-app/src/lib/helpers/stellar/tokenUtils.ts
